[k8s][feat] Implement tag update, tag delete and resource delete
