+++ plugins/k8s/resoto_plugin_k8s/collector.py
@@ -1,9 +1,7 @@
 import logging
 
-from kubernetes.client import Configuration
-
-from resoto_plugin_k8s.client import K8sClient
-from resoto_plugin_k8s.config import K8sConfig
+from resoto_plugin_k8s.base import K8sClient
+from resoto_plugin_k8s.base import K8sConfig
 from resoto_plugin_k8s.resources import (
     KubernetesCluster,
     all_k8s_resources_by_k8s_name,
@@ -14,7 +12,7 @@ from resoto_plugin_k8s.resources import (
 from resotolib.baseresources import EdgeType
 from resotolib.graph import Graph
 
-log = logging.getLogger("resoto." + __name__)
+log = logging.getLogger("resoto.plugins.k8s")
 
 
 class KubernetesCollector:
@@ -29,12 +27,8 @@ class KubernetesCollector:
     containing all K8S resources.
     """
 
-    def __init__(
-        self, k8s_config: K8sConfig, cluster_id: str, cluster_config: Configuration, client: K8sClient
-    ) -> None:
+    def __init__(self, k8s_config: K8sConfig, client: K8sClient) -> None:
         self.k8s_config = k8s_config
-        self.cluster_id = cluster_id
-        self.config = cluster_config
         self.client = client
         self.graph = Graph(root=self.cluster())
         self.builder = GraphBuilder(self.graph)
@@ -42,14 +36,14 @@ class KubernetesCollector:
     def cluster(self) -> KubernetesCluster:
         v = self.client.version()
         return KubernetesCluster(
-            id=self.cluster_id,
-            name=self.config.host,
+            id=self.client.cluster_id,
+            name=self.client.host,
             cluster_info=KubernetesClusterInfo(v.get("major", ""), v.get("minor", ""), v.get("platform", "")),
         )
 
     def collect(self) -> None:
         # collect all resources
-        for resource in self.client.apis():
+        for resource in self.client.apis:
             known = all_k8s_resources_by_k8s_name.get(resource.kind)
             if known and self.k8s_config.is_allowed(resource.kind):
                 for res, source in self.client.list_resources(resource, known):
